requirements.txt
dotenv
streamlit
google.generativeai
speech_recognition
pyttsx3
folium
geopy.geocoders
pil